+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## 0.6.1
+* Added `getClosestEventDate` and `getPoolDefinitionsAsDates` to `ticket-pool-parser`
+* Changed some `ticket-pool-parser` tests
+
 ## 0.6.0
 
 * Added `ticket-pool-parser` utils